+++ firebase-firestore/src/main/java/com/google/firebase/firestore/local/LocalDocumentsView.java
@@ -363,10 +363,10 @@ class LocalDocumentsView {
 
   private ImmutableSortedMap<DocumentKey, Document> getDocumentsMatchingCollectionQuery(
       Query query, IndexOffset offset) {
-    Map<DocumentKey, MutableDocument> remoteDocuments =
-        remoteDocumentCache.getAll(query.getPath(), offset);
     Map<DocumentKey, Overlay> overlays =
         documentOverlayCache.getOverlays(query.getPath(), offset.getLargestBatchId());
+    Map<DocumentKey, MutableDocument> remoteDocuments =
+        remoteDocumentCache.getDocumentsMatchingQuery(query, offset, overlays.keySet());
 
     // As documents might match the query because of their overlay we need to include documents
     // for all overlays in the initial document set.