+++ firebase-perf/src/test/java/com/google/firebase/perf/metrics/AppStartTraceTest.java
@@ -25,6 +25,7 @@ import static org.mockito.MockitoAnnotations.initMocks;
 import static org.robolectric.Shadows.shadowOf;
 
 import android.app.Activity;
+import android.content.Context;
 import android.os.Bundle;
 import android.os.Looper;
 import android.os.Process;
@@ -70,6 +71,7 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
   @Mock private Activity activity2;
   @Mock private Bundle bundle;
 
+  private final Context appContext = ApplicationProvider.getApplicationContext();
   private ArgumentCaptor<TraceMetric> traceArgumentCaptor;
 
   // a mocked current wall-clock time in microseconds.
@@ -103,6 +105,7 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
     FakeScheduledExecutorService fakeExecutorService = new FakeScheduledExecutorService();
     AppStartTrace trace =
         new AppStartTrace(transportManager, clock, configResolver, fakeExecutorService);
+    trace.registerActivityLifecycleCallbacks(appContext);
     // first activity goes through onCreate()->onStart()->onResume() state change.
     currentTime = 1;
     trace.onActivityCreated(activity1, bundle);
@@ -176,6 +179,7 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
     FakeScheduledExecutorService fakeExecutorService = new FakeScheduledExecutorService();
     AppStartTrace trace =
         new AppStartTrace(transportManager, clock, configResolver, fakeExecutorService);
+    trace.registerActivityLifecycleCallbacks(appContext);
     // first activity onCreate()
     currentTime = 1;
     trace.onActivityCreated(activity1, bundle);
@@ -212,6 +216,7 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
     FakeScheduledExecutorService fakeExecutorService = new FakeScheduledExecutorService();
     AppStartTrace trace =
         new AppStartTrace(transportManager, clock, configResolver, fakeExecutorService);
+    trace.registerActivityLifecycleCallbacks(appContext);
     // Delays activity creation after 1 minute from app start time.
     currentTime =
         TimeUnit.MILLISECONDS.toMicros(SystemClock.elapsedRealtime())
@@ -258,25 +263,28 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
   public void timeToInitialDisplay_isLogged() {
     // Test setup
     when(clock.getTime()).thenCallRealMethod(); // Use robolectric shadows to manipulate time
-    View testView = new View(ApplicationProvider.getApplicationContext());
+    View testView = new View(appContext);
     when(activity1.findViewById(android.R.id.content)).thenReturn(testView);
+    when(activity2.findViewById(android.R.id.content)).thenReturn(testView);
     when(configResolver.getIsExperimentTTIDEnabled()).thenReturn(true);
     FakeScheduledExecutorService fakeExecutorService = new FakeScheduledExecutorService();
     AppStartTrace trace =
         new AppStartTrace(transportManager, clock, configResolver, fakeExecutorService);
-
+    trace.registerActivityLifecycleCallbacks(appContext);
     // Simulate resume and manually stepping time forward
     ShadowSystemClock.advanceBy(Duration.ofMillis(1000));
     long resumeTime = TimeUnit.NANOSECONDS.toMicros(SystemClock.elapsedRealtimeNanos());
     trace.onActivityCreated(activity1, bundle);
     trace.onActivityStarted(activity1);
     trace.onActivityResumed(activity1);
-    // Experiment: simulate backgrounding before draw
+    // Experiment: simulate new activity before draw
+    trace.onActivityStarted(activity2);
+    trace.onActivityResumed(activity2);
     trace.onActivityPaused(activity1);
     trace.onActivityStopped(activity1);
-    trace.onActivityStarted(activity1);
-    trace.onActivityResumed(activity1);
+    trace.onActivityDestroyed(activity1);
     fakeExecutorService.runAll();
+    // only the old original _as should have been sent
     verify(transportManager, times(1))
         .log(isA(TraceMetric.class), isA(ApplicationProcessState.class));
 
@@ -296,6 +304,6 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
     assertThat(ttid.getName()).isEqualTo("_experiment_app_start_ttid");
     assertThat(ttid.getDurationUs()).isNotEqualTo(resumeTime - appStartTime);
     assertThat(ttid.getDurationUs()).isEqualTo(drawTime - appStartTime);
-    assertThat(ttid.getSubtracesCount()).isEqualTo(6);
+    assertThat(ttid.getSubtracesCount()).isEqualTo(3);
   }
 }