+++ firebase-installations/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Unreleased
 
+# 17.1.3
+*   [changed] Internal changes to improve startup time
+
+# 17.1.2
+*   [fixed] Updated `firebase-common` to its latest version (v20.3.0) to fix an issue that was 
+    causing a nondeterministic crash on startup.
+
 # 17.1.0
 *   [changed] Internal changes to ensure functionality alignment with other
     SDK releases. For more details, refer to the